+++ gymnax/utils/__init__.py
@@ -1,8 +1,10 @@
 from gymnax.utils.state_translate import np_state_to_jax
 from gymnax.utils.test_helpers import (assert_correct_state,
-                                       assert_correct_transit)
+                                       assert_correct_transit,
+                                       minatar_action_map)
 
 
 __all__ = ["np_state_to_jax",
            "assert_correct_state",
-           "assert_correct_transit"]
+           "assert_correct_transit",
+           "minatar_action_map"]