+++ tests/classic_control/test_gym_env.py
@@ -9,9 +9,7 @@ from gymnax.utils import (np_state_to_jax,
                           assert_correct_transit,
                           assert_correct_state)
 
-
-num_episodes, num_steps = 10, 150
-tolerance = 1e-04
+num_episodes, num_steps, tolerance = 10, 150, 1e-04
 
 
 def test_step(env_name):